Move GracefulKiller to KartonServiceBase
